Introducing radix-engine-store-interface + narrowing dependencies + fixing imports.
