+++ radix-engine-queries/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [dependencies]
 sbor = { path = "../sbor", default-features = false }
 radix-engine = { path = "../radix-engine", default-features = false }
-radix-engine-stores = { path = "../radix-engine-stores", default-features = false }
+radix-engine-store-interface = { path = "../radix-engine-store-interface", default-features = false }
 radix-engine-constants = { path = "../radix-engine-constants", default-features = false }
 radix-engine-interface = { path = "../radix-engine-interface", default-features = false }
 transaction = { path = "../transaction", default-features = false }
@@ -17,8 +17,8 @@ hex = { version = "0.4.3", default-features = false }
 [features]
 # You should enable either `std` or `alloc`
 default = ["std"]
-std = ["hex/std", "sbor/std", "transaction/std", "radix-engine-interface/std", "radix-engine-stores/std", "utils/std"]
-alloc = ["hex/alloc", "sbor/alloc", "transaction/alloc", "radix-engine-interface/alloc", "radix-engine-stores/alloc", "utils/alloc"]
+std = ["hex/std", "sbor/std", "transaction/std", "radix-engine-interface/std", "radix-engine-store-interface/std", "utils/std"]
+alloc = ["hex/alloc", "sbor/alloc", "transaction/alloc", "radix-engine-interface/alloc", "radix-engine-store-interface/alloc", "utils/alloc"]
 
 # Ref: https://bheisler.github.io/criterion.rs/book/faq.html#cargo-bench-gives-unrecognized-option-errors-for-valid-command-line-options
 [lib]
