+++ radix-engine-queries/src/query/traverse.rs
@@ -1,4 +1,5 @@
 use radix_engine::system::node_modules::type_info::TypeInfoSubstate;
+use radix_engine::track::db_key_mapper::{JmtMapper, MappedSubstateDatabase};
 use radix_engine_interface::blueprints::resource::{
     LiquidNonFungibleVault, FUNGIBLE_VAULT_BLUEPRINT, NON_FUNGIBLE_VAULT_BLUEPRINT,
 };
@@ -10,7 +11,7 @@ use radix_engine_interface::types::{
     ObjectInfo, ResourceAddress, SysModuleId, TypeInfoOffset,
 };
 use radix_engine_interface::{blueprints::resource::LiquidFungibleResource, types::NodeId};
-use radix_engine_stores::{interface::SubstateDatabase, jmt_support::JmtMapper};
+use radix_engine_store_interface::interface::SubstateDatabase;
 use sbor::rust::prelude::*;
 
 pub struct StateTreeTraverser<'s, 'v, S: SubstateDatabase, V: StateTreeVisitor> {
