+++ radix-engine-store-interface/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "radix-engine-stores"
+name = "radix-engine-store-interface"
 version = "0.10.0"
 edition = "2021"
 
@@ -8,17 +8,14 @@ sbor = { path = "../sbor", default-features = false }
 utils = { path = "../utils", default-features = false }
 radix-engine-common = { path = "../radix-engine-common", default-features = false }
 radix-engine-derive = { path = "../radix-engine-derive", default-features = false }
-radix-engine-store-interface = { path = "../radix-engine-store-interface", default-features = false }
-rocksdb = { version = "0.19.0", optional = true }
 itertools = { version = "0.10.3", default-features = false }
 hex = { version = "0.4.3", default-features = false }
 
 [features]
 default = ["std"]
-std = ["hex/std", "sbor/std", "utils/std", "radix-engine-common/std", "radix-engine-derive/std", "radix-engine-store-interface/std", "itertools/use_std"]
-alloc = ["hex/alloc", "sbor/alloc", "utils/alloc", "radix-engine-common/alloc", "radix-engine-derive/alloc", "radix-engine-store-interface/alloc", "itertools/use_alloc"]
+std = ["hex/std", "utils/std", "sbor/std", "radix-engine-derive/std", "radix-engine-common/std", "itertools/use_std"]
+alloc = ["hex/alloc", "utils/alloc", "sbor/alloc", "radix-engine-derive/alloc", "radix-engine-common/alloc", "itertools/use_alloc"]
 
-rocksdb = ["dep:rocksdb"]
 
 # Ref: https://bheisler.github.io/criterion.rs/book/faq.html#cargo-bench-gives-unrecognized-option-errors-for-valid-command-line-options
 [lib]
