+++ radix-engine-store-interface/src/lib.rs
@@ -5,9 +5,4 @@ compile_error!("Either feature `std` or `alloc` must be enabled for this crate."
 #[cfg(all(feature = "std", feature = "alloc"))]
 compile_error!("Feature `std` and `alloc` can't be enabled at the same time.");
 
-mod utils;
-
-pub mod hash_tree;
-pub mod memory_db;
-#[cfg(feature = "rocksdb")]
-pub mod rocks_db;
+pub mod interface;
