+++ radix-engine-stores/src/hash_tree/test.rs
@@ -5,27 +5,16 @@ use crate::hash_tree::tree_store::{
 };
 use crate::hash_tree::types::{LeafKey, NodeKey};
 use crate::hash_tree::{put_at_next_version, DbId, SubstateHashChange};
-use crate::interface::DatabaseMapper;
-use crate::jmt_support::JmtMapper;
 use itertools::Itertools;
-use radix_engine_interface::crypto::{hash, Hash};
-use radix_engine_interface::data::scrypto::{scrypto_decode, scrypto_encode};
-use radix_engine_interface::types::{ModuleId, NodeId, PackageAddress, SubstateKey, SysModuleId};
-use sbor::rust::collections::{hashmap, hashset, HashMap, HashSet};
+use radix_engine_common::crypto::{hash, Hash};
+use radix_engine_common::data::scrypto::{scrypto_decode, scrypto_encode};
+use utils::rust::collections::{hashmap, hashset, HashMap, HashSet};
 
 #[test]
 fn hash_of_next_version_differs_when_value_changed() {
     let mut store = TypedInMemoryTreeStore::new();
-    let hash_v1 = put_at_next_version(
-        &mut store,
-        None,
-        vec![change(1, SysModuleId::Object, 2, Some(30))],
-    );
-    let hash_v2 = put_at_next_version(
-        &mut store,
-        Some(1),
-        vec![change(1, SysModuleId::Object, 2, Some(70))],
-    );
+    let hash_v1 = put_at_next_version(&mut store, None, vec![change(1, 2, Some(30))]);
+    let hash_v2 = put_at_next_version(&mut store, Some(1), vec![change(1, 2, Some(70))]);
     assert_ne!(hash_v1, hash_v2);
 }
 
@@ -35,16 +24,9 @@ fn hash_of_next_version_same_when_write_repeated() {
     let hash_v1 = put_at_next_version(
         &mut store,
         None,
-        vec![
-            change(4, SysModuleId::Object, 6, Some(30)),
-            change(3, SysModuleId::Object, 9, Some(40)),
-        ],
+        vec![change(4, 6, Some(30)), change(3, 9, Some(40))],
     );
-    let hash_v2 = put_at_next_version(
-        &mut store,
-        Some(1),
-        vec![change(4, SysModuleId::Object, 6, Some(30))],
-    );
+    let hash_v2 = put_at_next_version(&mut store, Some(1), vec![change(4, 6, Some(30))]);
     assert_eq!(hash_v1, hash_v2);
 }
 
@@ -54,47 +36,29 @@ fn hash_of_next_version_same_when_write_empty() {
     let hash_v1 = put_at_next_version(
         &mut store,
         None,
-        vec![
-            change(1, SysModuleId::Object, 2, Some(30)),
-            change(3, SysModuleId::Object, 1, Some(40)),
-        ],
+        vec![change(1, 2, Some(30)), change(3, 1, Some(40))],
     );
     let hash_v2 = put_at_next_version(&mut store, Some(1), vec![]);
     assert_eq!(hash_v1, hash_v2);
 }
 
 #[test]
-fn hash_of_next_version_differs_when_substate_added() {
+fn hash_of_next_version_differs_when_entry_added() {
     let mut store = TypedInMemoryTreeStore::new();
-    let hash_v1 = put_at_next_version(
-        &mut store,
-        None,
-        vec![change(1, SysModuleId::Object, 2, Some(30))],
-    );
-    let hash_v2 = put_at_next_version(
-        &mut store,
-        Some(1),
-        vec![change(1, SysModuleId::Object, 8, Some(30))],
-    );
+    let hash_v1 = put_at_next_version(&mut store, None, vec![change(1, 2, Some(30))]);
+    let hash_v2 = put_at_next_version(&mut store, Some(1), vec![change(1, 8, Some(30))]);
     assert_ne!(hash_v1, hash_v2);
 }
 
 #[test]
-fn hash_of_next_version_differs_when_substate_removed() {
+fn hash_of_next_version_differs_when_entry_removed() {
     let mut store = TypedInMemoryTreeStore::new();
     let hash_v1 = put_at_next_version(
         &mut store,
         None,
-        vec![
-            change(1, SysModuleId::Object, 2, Some(30)),
-            change(4, SysModuleId::Object, 3, Some(20)),
-        ],
+        vec![change(1, 2, Some(30)), change(4, 3, Some(20))],
     );
-    let hash_v2 = put_at_next_version(
-        &mut store,
-        Some(1),
-        vec![change(1, SysModuleId::Object, 2, None)],
-    );
+    let hash_v2 = put_at_next_version(&mut store, Some(1), vec![change(1, 2, None)]);
     assert_ne!(hash_v1, hash_v2);
 }
 
@@ -104,94 +68,58 @@ fn hash_returns_to_same_when_previous_state_restored() {
     let hash_v1 = put_at_next_version(
         &mut store,
         None,
-        vec![
-            change(1, SysModuleId::Object, 2, Some(30)),
-            change(3, SysModuleId::Object, 1, Some(40)),
-        ],
+        vec![change(1, 2, Some(30)), change(3, 1, Some(40))],
     );
     put_at_next_version(
         &mut store,
         Some(1),
         vec![
-            change(1, SysModuleId::Object, 2, Some(90)),
-            change(3, SysModuleId::Object, 1, None),
-            change(1, SysModuleId::Object, 5, Some(10)),
+            change(1, 2, Some(90)),
+            change(3, 1, None),
+            change(1, 5, Some(10)),
         ],
     );
     let hash_v3 = put_at_next_version(
         &mut store,
         Some(1),
         vec![
-            change(1, SysModuleId::Object, 2, Some(30)),
-            change(3, SysModuleId::Object, 1, Some(40)),
-            change(1, SysModuleId::Object, 5, None),
+            change(1, 2, Some(30)),
+            change(3, 1, Some(40)),
+            change(1, 5, None),
         ],
     );
     assert_eq!(hash_v1, hash_v3);
 }
 
 #[test]
-fn hash_differs_when_states_only_differ_by_node_id() {
+fn hash_differs_when_states_only_differ_by_partition_key() {
     let mut store_1 = TypedInMemoryTreeStore::new();
-    let hash_1 = put_at_next_version(
-        &mut store_1,
-        None,
-        vec![change(1, SysModuleId::Object, 3, Some(30))],
-    );
+    let hash_1 = put_at_next_version(&mut store_1, None, vec![change(1, 3, Some(30))]);
     let mut store_2 = TypedInMemoryTreeStore::new();
-    let hash_2 = put_at_next_version(
-        &mut store_2,
-        None,
-        vec![change(2, SysModuleId::Object, 3, Some(30))],
-    );
+    let hash_2 = put_at_next_version(&mut store_2, None, vec![change(2, 3, Some(30))]);
     assert_ne!(hash_1, hash_2);
 }
 
 #[test]
-fn hash_differs_when_states_only_differ_by_module_id() {
+fn hash_differs_when_states_only_differ_by_sort_key() {
     let mut store_1 = TypedInMemoryTreeStore::new();
-    let hash_1 = put_at_next_version(
-        &mut store_1,
-        None,
-        vec![change(1, SysModuleId::Metadata, 3, Some(30))],
-    );
+    let hash_1 = put_at_next_version(&mut store_1, None, vec![change(1, 2, Some(30))]);
     let mut store_2 = TypedInMemoryTreeStore::new();
-    let hash_2 = put_at_next_version(
-        &mut store_2,
-        None,
-        vec![change(1, SysModuleId::AccessRules, 3, Some(30))],
-    );
+    let hash_2 = put_at_next_version(&mut store_2, None, vec![change(1, 3, Some(30))]);
     assert_ne!(hash_1, hash_2);
 }
 
 #[test]
-fn hash_differs_when_states_only_differ_by_offset() {
-    let mut store_1 = TypedInMemoryTreeStore::new();
-    let hash_1 = put_at_next_version(
-        &mut store_1,
-        None,
-        vec![change(1, SysModuleId::Object, 2, Some(30))],
-    );
-    let mut store_2 = TypedInMemoryTreeStore::new();
-    let hash_2 = put_at_next_version(
-        &mut store_2,
-        None,
-        vec![change(1, SysModuleId::Object, 3, Some(30))],
-    );
-    assert_ne!(hash_1, hash_2);
-}
-
-#[test]
-fn hash_of_different_re_node_nested_trees_is_same_when_contained_substates_are_same() {
+fn hash_of_different_partition_nested_trees_is_same_when_contained_entries_are_same() {
     let mut store = TypedInMemoryTreeStore::new();
     put_at_next_version(
         &mut store,
         None,
         vec![
-            change(1, SysModuleId::Metadata, 2, Some(30)),
-            change(1, SysModuleId::Metadata, 9, Some(40)),
-            change(7, SysModuleId::AccessRules, 2, Some(30)),
-            change(7, SysModuleId::AccessRules, 9, Some(40)),
+            change(1, 2, Some(30)),
+            change(1, 9, Some(40)),
+            change(7, 2, Some(30)),
+            change(7, 9, Some(40)),
         ],
     );
 
@@ -208,7 +136,7 @@ fn hash_of_different_re_node_nested_trees_is_same_when_contained_substates_are_s
 }
 
 #[test]
-fn db_index_and_key_are_used_directly_for_node_nibble_path() {
+fn partition_key_and_sort_key_are_used_directly_for_node_nibble_path() {
     let mut store = TypedInMemoryTreeStore::new();
 
     put_at_next_version(
@@ -279,7 +207,7 @@ fn db_index_and_key_are_used_directly_for_node_nibble_path() {
 }
 
 #[test]
-fn deletes_re_node_layer_leaf_when_all_its_substates_deleted() {
+fn deletes_partition_layer_leaf_when_all_its_entries_deleted() {
     fn count_current_re_node_leafs(store: &TypedInMemoryTreeStore) -> usize {
         store
             .root_tree_nodes
@@ -294,26 +222,19 @@ fn deletes_re_node_layer_leaf_when_all_its_substates_deleted() {
         &mut store,
         None,
         vec![
-            change(1, SysModuleId::Metadata, 2, Some(30)),
-            change(1, SysModuleId::Metadata, 9, Some(40)),
-            change(1, SysModuleId::AccessRules, 3, Some(30)),
+            change(1, 2, Some(30)),
+            change(1, 9, Some(40)),
+            change(2, 3, Some(30)),
         ],
     );
     assert_eq!(count_current_re_node_leafs(&store), 2);
     put_at_next_version(
         &mut store,
         Some(1),
-        vec![
-            change(1, SysModuleId::Metadata, 2, None),
-            change(1, SysModuleId::Metadata, 9, None),
-        ],
+        vec![change(1, 2, None), change(1, 9, None)],
     );
     assert_eq!(count_current_re_node_leafs(&store), 1);
-    put_at_next_version(
-        &mut store,
-        Some(2),
-        vec![change(1, SysModuleId::AccessRules, 3, None)],
-    );
+    put_at_next_version(&mut store, Some(2), vec![change(2, 3, None)]);
     assert_eq!(count_current_re_node_leafs(&store), 0);
 }
 
@@ -322,38 +243,18 @@ fn supports_empty_state() {
     let mut store = TypedInMemoryTreeStore::new();
     let hash_v1 = put_at_next_version(&mut store, None, vec![]);
     assert_eq!(hash_v1, SPARSE_MERKLE_PLACEHOLDER_HASH);
-    let hash_v2 = put_at_next_version(
-        &mut store,
-        Some(1),
-        vec![change(1, SysModuleId::Object, 2, Some(30))],
-    );
+    let hash_v2 = put_at_next_version(&mut store, Some(1), vec![change(1, 2, Some(30))]);
     assert_ne!(hash_v2, SPARSE_MERKLE_PLACEHOLDER_HASH);
-    let hash_v3 = put_at_next_version(
-        &mut store,
-        Some(2),
-        vec![change(1, SysModuleId::Object, 2, None)],
-    );
+    let hash_v3 = put_at_next_version(&mut store, Some(2), vec![change(1, 2, None)]);
     assert_eq!(hash_v3, SPARSE_MERKLE_PLACEHOLDER_HASH);
 }
 
 #[test]
 fn records_stale_tree_node_keys() {
     let mut store = TypedInMemoryTreeStore::new();
-    put_at_next_version(
-        &mut store,
-        None,
-        vec![change(4, SysModuleId::Object, 6, Some(30))],
-    );
-    put_at_next_version(
-        &mut store,
-        Some(1),
-        vec![change(3, SysModuleId::Object, 9, Some(70))],
-    );
-    put_at_next_version(
-        &mut store,
-        Some(2),
-        vec![change(3, SysModuleId::Object, 9, Some(80))],
-    );
+    put_at_next_version(&mut store, None, vec![change(4, 6, Some(30))]);
+    put_at_next_version(&mut store, Some(1), vec![change(3, 9, Some(70))]);
+    put_at_next_version(&mut store, Some(2), vec![change(3, 9, Some(80))]);
     let stale_versions = store
         .stale_key_buffer
         .iter()
@@ -414,17 +315,9 @@ fn sbor_decodes_what_was_encoded() {
 #[test]
 fn serialized_keys_are_strictly_increasing() {
     let mut store = SerializedInMemoryTreeStore::new();
-    put_at_next_version(
-        &mut store,
-        None,
-        vec![change(3, SysModuleId::Object, 4, Some(90))],
-    );
+    put_at_next_version(&mut store, None, vec![change(3, 4, Some(90))]);
     let previous_key = store.memory.keys().collect_vec()[0].clone();
-    put_at_next_version(
-        &mut store,
-        Some(1),
-        vec![change(1, SysModuleId::Object, 2, Some(80))],
-    );
+    put_at_next_version(&mut store, Some(1), vec![change(1, 2, Some(80))]);
     let next_key = store
         .memory
         .keys()
@@ -435,17 +328,14 @@ fn serialized_keys_are_strictly_increasing() {
 }
 
 fn change(
-    node_id_seed: u8,
-    module_id: SysModuleId,
-    substate_offset_seed: u8,
+    partition_key_seed: u8,
+    sort_key_seed: u8,
     value_hash_seed: Option<u8>,
 ) -> SubstateHashChange {
-    let fake_pkg_address = unsafe { PackageAddress::new_unchecked([node_id_seed; NodeId::LENGTH]) };
-    let fake_kvs_entry_id = vec![substate_offset_seed; substate_offset_seed as usize];
     SubstateHashChange::new(
         DbId::new(
-            JmtMapper::map_to_db_index(&NodeId(fake_pkg_address.into()), ModuleId(module_id as u8)),
-            JmtMapper::map_to_db_key(&SubstateKey::Map(fake_kvs_entry_id)),
+            vec![partition_key_seed; Hash::LENGTH],
+            vec![sort_key_seed; sort_key_seed as usize],
         ),
         value_hash_seed.map(|value_seed| value_hash(value_seed)),
     )
