+++ radix-engine-tests/Cargo.toml
@@ -9,6 +9,7 @@ radix-engine = { path = "../radix-engine", default-features = false }
 radix-engine-constants = { path = "../radix-engine-constants" }
 radix-engine-interface = { path = "../radix-engine-interface", default-features = false }
 radix-engine-common = { path = "../radix-engine-common", default-features = false }
+radix-engine-store-interface = { path = "../radix-engine-store-interface", default-features = false }
 radix-engine-stores = { path = "../radix-engine-stores", default-features = false }
 radix-engine-queries = { path = "../radix-engine-queries", default-features = false }
 transaction = { path = "../transaction", default-features = false }
@@ -61,8 +62,8 @@ harness = false
 [features]
 # You should enable either `std` or `alloc`
 default = ["std"]
-std = ["sbor/std", "transaction/std", "radix-engine/std", "radix-engine-interface/std", "radix-engine-stores/std", "radix-engine-queries/std", "utils/std", "scrypto/std", "scrypto-unit/std"]
-alloc = ["sbor/alloc", "transaction/alloc", "radix-engine/alloc", "radix-engine-interface/alloc", "radix-engine-stores/alloc", "radix-engine-queries/alloc", "utils/alloc", "scrypto/alloc", "scrypto-unit/alloc"]
+std = ["sbor/std", "transaction/std", "radix-engine/std", "radix-engine-interface/std", "radix-engine-stores/std", "radix-engine-store-interface/std", "radix-engine-queries/std", "utils/std", "scrypto/std", "scrypto-unit/std"]
+alloc = ["sbor/alloc", "transaction/alloc", "radix-engine/alloc", "radix-engine-interface/alloc", "radix-engine-stores/alloc", "radix-engine-store-interface/alloc", "radix-engine-queries/alloc", "utils/alloc", "scrypto/alloc", "scrypto-unit/alloc"]
 wasmer = ["radix-engine/wasmer"]
 cpu_ram_metrics = ["radix-engine/cpu_ram_metrics"]
 flamegraph = []
