+++ radix-engine-tests/tests/fee.rs
@@ -2,10 +2,10 @@ use radix_engine::blueprints::resource::WorktopError;
 use radix_engine::errors::{ApplicationError, CallFrameError, KernelError};
 use radix_engine::errors::{RejectionError, RuntimeError};
 use radix_engine::kernel::call_frame::LockSubstateError;
+use radix_engine::track::interface::AcquireLockError;
 use radix_engine::transaction::TransactionReceipt;
 use radix_engine::types::*;
 use radix_engine_interface::blueprints::resource::FromPublicKey;
-use radix_engine_stores::interface::AcquireLockError;
 use scrypto_unit::*;
 use transaction::builder::ManifestBuilder;
 use transaction::model::*;
