+++ radix-engine/src/system/module.rs
@@ -5,9 +5,9 @@ use crate::kernel::kernel_api::KernelApi;
 use crate::kernel::kernel_api::KernelInvocation;
 use crate::kernel::kernel_callback_api::KernelCallbackObject;
 use crate::system::system_callback::SystemInvocation;
+use crate::track::interface::NodeSubstates;
 use crate::types::*;
 use radix_engine_interface::api::substate_lock_api::LockFlags;
-use radix_engine_stores::interface::NodeSubstates;
 
 pub trait SystemModule<M: KernelCallbackObject> {
     //======================
