+++ radix-engine/src/system/system.rs
@@ -15,6 +15,7 @@ use crate::system::system_callback_api::SystemCallbackObject;
 use crate::system::system_modules::costing::FIXED_LOW_FEE;
 use crate::system::system_modules::events::EventError;
 use crate::system::system_modules::execution_trace::{BucketSnapshot, ProofSnapshot};
+use crate::track::interface::NodeSubstates;
 use crate::types::*;
 use radix_engine_interface::api::index_api::ClientIndexApi;
 use radix_engine_interface::api::key_value_store_api::ClientKeyValueStoreApi;
@@ -33,7 +34,6 @@ use radix_engine_interface::blueprints::identity::*;
 use radix_engine_interface::blueprints::package::*;
 use radix_engine_interface::blueprints::resource::*;
 use radix_engine_interface::schema::KeyValueStoreSchema;
-use radix_engine_stores::interface::NodeSubstates;
 use resources_tracker_macro::trace_resources;
 use sbor::rust::string::ToString;
 use sbor::rust::vec::Vec;
