+++ radix-engine/src/track/interface.rs
@@ -1,41 +1,6 @@
-/*
-
-High-level Abstraction
-
-+-------------------------+
-|       Radix Engine      |
-|----> SubstateStore <----|
-|          Track          |
-|---> SubstateDatabase <--|
-|         Database        |
-+-------------------------+
-
-*/
-
-use radix_engine_interface::api::LockFlags;
-use radix_engine_interface::data::scrypto::{scrypto_decode, ScryptoDecode};
+use crate::types::*;
+use radix_engine_interface::api::substate_lock_api::LockFlags;
 use radix_engine_interface::types::*;
-use radix_engine_interface::*;
-use sbor::rust::prelude::*;
-
-pub fn encode_substate_id(index_id: &Vec<u8>, db_key: &Vec<u8>) -> Vec<u8> {
-    let mut buffer = Vec::new();
-    buffer.extend(index_id);
-    buffer.extend(db_key); // Length is marked by EOF
-    buffer
-}
-
-// TODO: Clean this interface up and move size of hash to a more appropriate interface
-pub fn decode_substate_id(slice: &[u8]) -> Option<(Vec<u8>, Vec<u8>)> {
-    if slice.len() >= 26 {
-        let index_id = slice[0..26].to_vec();
-        let key = slice[26 + 1..].to_vec();
-
-        return Some((index_id, key));
-    }
-
-    return None;
-}
 
 /// Error when acquiring a lock.
 #[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor)]
@@ -157,74 +122,4 @@ pub trait SubstateStore {
     /// - If the lock handle is invalid;
     /// - If the lock handle is not associated with WRITE permission
     fn update_substate(&mut self, handle: u32, substate_value: IndexedScryptoValue);
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor)]
-pub struct StateUpdates {
-    pub database_updates: DatabaseUpdates,
-    pub system_updates: SystemUpdates,
-}
-
-pub type DatabaseUpdates = IndexMap<Vec<u8>, IndexMap<Vec<u8>, DatabaseUpdate>>;
-pub type SystemUpdates = IndexMap<(NodeId, ModuleId), IndexMap<SubstateKey, DatabaseUpdate>>;
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor)]
-pub enum DatabaseUpdate {
-    Set(Vec<u8>),
-    Delete,
-}
-
-pub trait DatabaseMapper {
-    fn map_to_db_index(node_id: &NodeId, module_id: ModuleId) -> Vec<u8>;
-    fn map_to_db_key(key: &SubstateKey) -> Vec<u8>;
-}
-
-/// Represents the interface between Track and a database vendor.
-pub trait SubstateDatabase {
-    /// Reads a substate of the given node module.
-    ///
-    /// [`Option::None`] is returned if missing.
-    fn get_substate(&self, index_id: &Vec<u8>, key: &Vec<u8>) -> Option<Vec<u8>>;
-
-    /// Returns a lexicographical sorted iterator over the substates of an index
-    fn list_substates(
-        &self,
-        index_id: &Vec<u8>,
-    ) -> Box<dyn Iterator<Item = (Vec<u8>, Vec<u8>)> + '_>;
-
-    /// Convenience method for database readers
-    fn get_mapped_substate<M: DatabaseMapper, D: ScryptoDecode>(
-        &self,
-        node_id: &NodeId,
-        module_id: ModuleId,
-        substate_key: &SubstateKey,
-    ) -> Option<D> {
-        self.get_substate(
-            &M::map_to_db_index(node_id, module_id),
-            &M::map_to_db_key(&substate_key),
-        )
-        .map(|buf| scrypto_decode(&buf).unwrap())
-    }
-
-    /// Convenience method for database readers
-    fn list_mapped_substates<M: DatabaseMapper>(
-        &self,
-        node_id: &NodeId,
-        module_id: ModuleId,
-    ) -> Box<dyn Iterator<Item = (Vec<u8>, Vec<u8>)> + '_> {
-        self.list_substates(&M::map_to_db_index(node_id, module_id))
-    }
-}
-
-/// Interface for committing changes into a substate database.
-pub trait CommittableSubstateDatabase {
-    /// Commits state changes to the database.
-    ///
-    /// An error is thrown in case of invalid module ID.
-    fn commit(&mut self, state_changes: &DatabaseUpdates);
-}
-
-/// Interface for listing nodes within a substate database.
-pub trait ListableSubstateDatabase {
-    fn list_nodes(&self) -> Vec<NodeId>;
 }
