+++ radix-engine/src/track/mod.rs
@@ -1,3 +1,5 @@
+pub mod db_key_mapper;
+pub mod interface;
 pub mod track;
 
 pub use track::*;
