+++ radix-engine/src/transaction/transaction_receipt.rs
@@ -5,6 +5,7 @@ use crate::system::system_modules::costing::FeeSummary;
 use crate::system::system_modules::execution_trace::{
     ExecutionTrace, ResourceChange, WorktopChange,
 };
+use crate::track::StateUpdates;
 use crate::types::*;
 use colored::*;
 use radix_engine_interface::address::AddressDisplayContext;
@@ -12,7 +13,6 @@ use radix_engine_interface::api::ObjectModuleId;
 use radix_engine_interface::blueprints::transaction_processor::InstructionOutput;
 use radix_engine_interface::data::scrypto::ScryptoDecode;
 use radix_engine_interface::types::*;
-use radix_engine_stores::interface::StateUpdates;
 use sbor::representations::*;
 use utils::ContextualDisplay;
