+++ scrypto-unit/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 radix-engine = { path = "../radix-engine", default-features = false }
 radix-engine-interface = { path = "../radix-engine-interface", default-features = false }
 radix-engine-constants = { path = "../radix-engine-constants" }
+radix-engine-store-interface = { path = "../radix-engine-store-interface", default-features = false }
 radix-engine-stores = { path = "../radix-engine-stores", default-features = false }
 radix-engine-queries = { path = "../radix-engine-queries", default-features = false }
 transaction = { path = "../transaction", default-features = false }
@@ -19,8 +20,8 @@ wabt = { version = "0.10.0" }
 
 [features]
 default = ["std"]
-std = ["scrypto/std", "scrypto/serde", "sbor/std", "radix-engine-interface/std", "radix-engine-stores/std", "radix-engine-queries/std", "radix-engine/std", "transaction/std", "utils/std"]
-alloc = ["scrypto/alloc", "scrypto/serde", "sbor/alloc", "radix-engine-interface/alloc", "radix-engine-stores/alloc", "radix-engine-queries/alloc", "radix-engine/alloc", "transaction/alloc", "utils/alloc"]
+std = ["scrypto/std", "scrypto/serde", "sbor/std", "radix-engine-interface/std", "radix-engine-stores/std", "radix-engine-store-interface/std", "radix-engine-queries/std", "radix-engine/std", "transaction/std", "utils/std"]
+alloc = ["scrypto/alloc", "scrypto/serde", "sbor/alloc", "radix-engine-interface/alloc", "radix-engine-stores/alloc", "radix-engine-store-interface/alloc", "radix-engine-queries/alloc", "radix-engine/alloc", "transaction/alloc", "utils/alloc"]
 resource_tracker = []
 
 [lib]
